requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
schedule==1.1.0
blinker==1.6.2
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
idna==3.4
secure-smtplib==0.1.1 